fix: potentially resume a thread that isn't even waiting at the SyncFifo
This bug needed a precise scheduling event, right after querying the suspended reader, but before the actual 'resume()'. Forthermore the reader's timeout must have been reached.
In this case it was possible for the said thread to leave the fifo, do something else and suspend again, but at a completely different location.