Fix core_cmFunc.h for Clang
The functions for accessing the FPSCR in the CMSIS header file use
inline assembly with the vfpcc register, which the Clang assembler does
not know. Therefore, compiling for the STM32F4 will fail. This also
means that clang tools like clang-tidy will complain with a
clang-diagnostic-error. Since the header file is public, this affects
users of Rodos too. Thankfully, CMSIS 6 fixed this issue a while ago by
using replacing the inline assembly with a builtin that is available on
both Clang and GCC. I just did the same. Open source repositories with
good commit messages FTW!