Don't set `VTOR` in `SystemInit()` if offset is 0
We don't need to set the vector table offset register to 0 in
`SystemInit()`, since that is the reset value anyway. The real reason
for not setting it, however, is that this removes the only need to build
two different versions of rodos: one for use with, one for use without a
bootloader. This makes life much easier.